test: address review comments
Move DialFileService upload/cache tests to dial_core_services_tests where
they belong; split TextFileToolsConfig into its own test file with a
meaningful name; fix top-level imports. Also type state_holder metadata
param as FileMetadata (dropping the Any default) and update the cache-hit
assertion to match the new contract.
