+++ src/tests/unit_tests/dial_core_services_tests/test_dial_file_service.py
@@ -1,6 +1,7 @@
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
+from aidial_client._exception import EtagMismatchError
 
 from quickapp.common.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 from quickapp.common.state_holder import StateHolder
@@ -12,9 +13,13 @@ DEFAULT_LIMIT = 10 * 1024 * 1024
 def _make_mock_dial_client(
     content_length: int = 100,
     file_content: bytes = b"file content",
+    upload_url: str = "files/bucket/generated-files/out.txt",
+    etag: str | None = "abc123",
 ) -> MagicMock:
     mock_metadata = MagicMock()
     mock_metadata.content_length = content_length
+    mock_metadata.etag = etag
+    mock_metadata.url = upload_url
 
     mock_download_result = MagicMock()
     mock_download_result.aget_content = AsyncMock(return_value=file_content)
@@ -22,6 +27,11 @@ def _make_mock_dial_client(
     mock_files = MagicMock()
     mock_files.get_metadata = AsyncMock(return_value=mock_metadata)
     mock_files.download = AsyncMock(return_value=mock_download_result)
+    mock_files.upload = AsyncMock(return_value=mock_metadata)
+
+    mock_bucket = MagicMock()
+    mock_bucket.appdata = None
+    mock_bucket.bucket = "mybucket"
 
     mock_resource_permissions = MagicMock()
     mock_resource_permissions.grant = AsyncMock(return_value=None)
@@ -29,6 +39,8 @@ def _make_mock_dial_client(
     mock_dial_client = MagicMock()
     mock_dial_client.files = mock_files
     mock_dial_client.resource_permissions = mock_resource_permissions
+    mock_dial_client.bucket = MagicMock()
+    mock_dial_client.bucket.get_raw = AsyncMock(return_value=mock_bucket)
     return mock_dial_client
 
 
@@ -67,14 +79,15 @@ class TestDownloadFile:
     @pytest.mark.asyncio
     async def test_cache_hit_returns_from_state(self):
         holder = StateHolder()
-        holder.store_file_data("files/cached.txt", b"cached content")
+        stored_metadata = MagicMock()
+        holder.store_file_data("files/cached.txt", b"cached content", stored_metadata)
         mock_dial_client = _make_mock_dial_client()
         svc = _make_service(dial_client=mock_dial_client, state_holder=holder)
 
         data, metadata = await svc.download_file("files/cached.txt")
 
         assert data == b"cached content"
-        assert metadata is None
+        assert metadata is stored_metadata
         mock_dial_client.files.get_metadata.assert_not_called()
         mock_dial_client.files.download.assert_not_called()
 
@@ -98,6 +111,80 @@ class TestDownloadFile:
 
         with pytest.raises(ValueError, match="exceeds the limit of 1024"):
             await svc.download_file("files/medium.bin")
+
+
+class TestUploadText:
+    @pytest.mark.asyncio
+    async def test_create_only_passes_if_none_match(self):
+        mock_client = _make_mock_dial_client(upload_url="files/b/generated-files/f.txt")
+        svc = _make_service(dial_client=mock_client)
+
+        result = await svc.upload_text(
+            url="files/b/generated-files/f.txt",
+            content="hello",
+            if_none_match="*",
+        )
+
+        assert result == "files/b/generated-files/f.txt"
+        mock_client.files.upload.assert_awaited_once()
+        call_kwargs = mock_client.files.upload.call_args
+        assert call_kwargs.kwargs.get("etag_if_none_match") == "*"
+        assert call_kwargs.kwargs.get("etag_if_match") is None
+
+    @pytest.mark.asyncio
+    async def test_update_passes_if_match(self):
+        mock_client = _make_mock_dial_client(upload_url="files/b/generated-files/f.txt")
+        svc = _make_service(dial_client=mock_client)
+
+        await svc.upload_text(
+            url="files/b/generated-files/f.txt",
+            content="new content",
+            if_match="etag-xyz",
+        )
+
+        call_kwargs = mock_client.files.upload.call_args
+        assert call_kwargs.kwargs.get("etag_if_match") == "etag-xyz"
+        assert call_kwargs.kwargs.get("etag_if_none_match") is None
+
+    @pytest.mark.asyncio
+    async def test_propagates_412_as_etag_mismatch_error(self):
+        mock_client = _make_mock_dial_client()
+        mock_client.files.upload = AsyncMock(side_effect=EtagMismatchError(message="412"))
+        svc = _make_service(dial_client=mock_client)
+
+        with pytest.raises(EtagMismatchError):
+            await svc.upload_text(url="files/b/f.txt", content="x", if_none_match="*")
+
+    @pytest.mark.asyncio
+    async def test_returns_confirmed_url(self):
+        mock_client = _make_mock_dial_client(upload_url="files/b/generated-files/notes.md")
+        svc = _make_service(dial_client=mock_client)
+
+        result = await svc.upload_text(
+            url="files/b/generated-files/notes.md",
+            content="# Notes\n",
+        )
+        assert result == "files/b/generated-files/notes.md"
+
+
+class TestInvalidateCache:
+    @pytest.mark.asyncio
+    async def test_subsequent_download_hits_network_after_invalidation(self):
+        file_bytes_v1 = b"version1"
+        file_bytes_v2 = b"version2"
+        mock_client = _make_mock_dial_client(file_content=file_bytes_v1)
+        svc = _make_service(dial_client=mock_client)
+
+        first, _ = await svc.download_file("files/b/f.txt")
+        assert first == file_bytes_v1
+
+        # Now invalidate and change what the mock returns
+        svc.invalidate_cache("files/b/f.txt")
+        mock_client.files.download.return_value.aget_content = AsyncMock(return_value=file_bytes_v2)
+        mock_client.files.get_metadata.return_value.content_length = 100
+
+        second, _ = await svc.download_file("files/b/f.txt")
+        assert second == file_bytes_v2
 
 
 class TestGrantPermissions:
